refactor(helix-tui): split `symbols.rs` into submodules
