+++ helix-tui/src/symbols/braille.rs
@@ -1,0 +1,8 @@
+pub const BLANK: u16 = 0x2800;
+
+pub const DOTS: [[u16; 2]; 4] = [
+    [0x0001, 0x0008],
+    [0x0002, 0x0010],
+    [0x0004, 0x0020],
+    [0x0040, 0x0080],
+];
